+++ examples/slint/src/maplibre.rs
@@ -3,25 +3,34 @@ use crate::MapAdapter;
 use maplibre_native::Height;
 use maplibre_native::ScreenCoordinate;
 use maplibre_native::Width;
+use maplibre_native::layers::SymbolAnchorType;
 use slint::ComponentHandle;
 use std::rc::Rc;
 mod headless;
-use headless::MapLibre;
+pub use headless::MapLibre;
 pub use headless::create_map;
+use image::ImageReader;
+use maplibre_native::Style;
+use maplibre_native::{GeoJsonSource, Latitude, Longitude, SymbolLayer};
 use maplibre_native::{X, Y};
 use std::cell::RefCell;
+use std::path::Path;
 
 pub fn init(ui: &MainWindow, map: &Rc<RefCell<MapLibre>>) {
+    loop {
+        map.borrow_mut().renderer().render_once();
+        if map.borrow_mut().style_loaded() {
+            style(map);
+            break;
+        }
+    }
+
     ui.on_map_size_changed({
         let map = Rc::downgrade(map);
         move |size| {
             let size =
                 maplibre_native::Size::new(Width(size.width as u32), Height(size.height as u32));
-            map.upgrade()
-                .unwrap()
-                .borrow_mut()
-                .renderer()
-                .set_map_size(size);
+            map.upgrade().unwrap().borrow_mut().renderer().set_map_size(size);
         }
     });
 
@@ -77,11 +86,29 @@ pub fn init(ui: &MainWindow, map: &Rc<RefCell<MapLibre>>) {
             const STEP: f64 = 1.2;
             let pos = ScreenCoordinate::new(X(x.into()), Y(y.into()));
             let scale = if delta > 0. { STEP } else { 1.0 / STEP };
-            map.upgrade()
-                .unwrap()
-                .borrow_mut()
-                .renderer()
-                .scale_by(scale, pos);
+            map.upgrade().unwrap().borrow_mut().renderer().scale_by(scale, pos);
         }
     });
+}
+
+fn style(map: &Rc<RefCell<MapLibre>>) {
+    let mut map_borrow = map.borrow_mut();
+    let mut style = Style::get_ref(map_borrow.renderer());
+
+    let image = ImageReader::open(
+        Path::new(env!("CARGO_MANIFEST_DIR")).join("ui").join("icons").join("Marker.png"),
+    )
+    .unwrap()
+    .decode()
+    .unwrap();
+    let image_id = style.add_image("The id", &image, true);
+
+    let mut geo_json_source = GeoJsonSource::new("geojsonsourceid");
+    geo_json_source.set_point(Latitude(52.67655), Longitude(13.28387));
+    let source_id = style.add_source(geo_json_source);
+
+    let layer = SymbolLayer::new("Layer id", &source_id);
+    layer.set_icon_image(&image_id);
+    layer.set_icon_anchor(SymbolAnchorType::Bottom);
+    style.add_layer(layer);
 }