+++ src/cpp/style.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace mln::bridge::style {
+
+}